% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send email via mailgun.}
\usage{
send_email(from, to, subject, text, attach = NULL, tag = NULL)
}
\arguments{
\item{from}{character. Either \code{"someone" --> someone@domain.com} or
\code{c("Somebody Amazing", "someone") --> "Somebody Amazing <someone@domain.com>"}}

\item{to}{character. Email addresses of recipients.}

\item{subject}{character. Email subject.}

\item{text}{character. Email text.}

\item{attach}{character. Filenames of your attachments.}

\item{tag}{character. Mailgun tags (I still don't really know what this is but why not).}
}
\description{
Must set \code{mailgunner.api_key} and \code{mailgunner.domain_name} options.
}
\seealso{
Official mailgun documentation
  \url{https://documentation.mailgun.com/quickstart-sending.html#send-via-api}
}

